refactor(api): extract shared dependency wiring into internal/bootstrap
The ~50-line recipe that builds services from the db repositories into
api.Dependencies was duplicated in cmd/ovumcy/main.go (buildDependencies)
and the api test helper (newTestHandlerDependencies), and had already
drifted once on the transactional DayService.

- Move the Dependencies aggregate and its workflow-service ports
  (Registration/Login/OIDC/RegisterPickupToken) into a new internal/apideps
  package, imported by both internal/api (via type aliases, so
  api.Dependencies and the port names are unchanged) and the composition
  root. apideps imports only services/models/security, never internal/db.
- Add internal/bootstrap.BuildDependencies as the single wiring recipe,
  parameterized by Options (attempt limits, OIDC config, optional OIDC
  stub). cmd/ovumcy and both api test helpers now call it.
- Production wiring is behaviourally identical. The api test helper now
  builds DayService transactionally (NewDayServiceWithTx), matching prod;
  previously it used the non-transactional constructor.

internal/api remains free of internal/db (verified via go list -deps).
go build/vet/test ./... all green.
